uievents/interface/click-event.htm

<!doctype html>
<html>
  <head>
    <title>Click event is a MouseEvent</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id='clicktarget'></div>
    <div id="log"></div>
    <script type="text/javascript">
    var clicktarget = document.querySelector("#clicktarget");
    var t = async_test("synthetic click event is a PointerEvent");
    clicktarget.addEventListener('click', t.step_func(function (e) {
      assert_equals(e.constructor, window.PointerEvent);
      assert_true(e instanceof window.PointerEvent);
      t.done();
    }));
    document.querySelector('#clicktarget').click();
    </script>
  </body>
</html>
